requirements.txt
pandas
numpy
matplotlib
seaborn
requests
reportlab
python-pptx
